Moved test from integration to unit tests
